Tela de Agendamento, fazendo o inserir

+++ src/startenglish/db/DAL/DALAgenda.java
@@ -12,63 +12,19 @@ public class DALAgenda {
     
     public boolean gravar(Agenda a){
         
-        String sql = "insert into agendaprova(professorid,alunoid,dataprova,horainicio,horafim,local,situacaoprova,nota) values($1,$2,$3,'$4','$5','$6',$7,'$8')";
-        sql = sql.replaceAll("$1", ""+a.getProfessor().getFunc().getID());
-        sql = sql.replaceAll("$2", ""+a.getAluno().getID());
-        sql = sql.replaceAll("$3", "" +a.getDataProva());
-        sql = sql.replaceAll("$4", "" + a.getHoraini());
-        sql = sql.replaceAll("$5", "" + a.getHorafim());
-        sql = sql.replaceAll("$6", "" + a.getLocal());
-        sql = sql.replaceAll("$7", "" + a.getSituacao_prova());
-        sql = sql.replaceAll("$8", "" + a.getNota());
+        String sql = "insert into agendaprova(professorid,alunoid,dataprova,horainicio,horafim,local,situacaoprova,nota) values(#1,#2,'#3','#4','#5','#6','#7',#8)";
+        sql = sql.replaceAll("#1", ""+a.getProfessor().getFunc().getID());
+        sql = sql.replaceAll("#2", ""+a.getAluno().getID());
+        sql = sql.replaceAll("#3", "" +a.getDataProva());
+        sql = sql.replaceAll("#4",  a.getHoraini());
+        sql = sql.replaceAll("#5", a.getHorafim());
+        sql = sql.replaceAll("#6",  a.getLocal());
+        sql = sql.replaceAll("#7", "" + a.getSituacao_prova());
+        sql = sql.replaceAll("#8", "" + a.getNota());
         
         return Banco.getCon().manipular(sql);
     }
-    
-     public boolean alterar(Agenda a){
-        
-        String sql = "update agendaprova set professorid = $1,alunoid = $2,dataprova = $3,horainicio = '$4', horafim = '$5' ,local = '$6',situacaoprova = '$7',nota = $8 where agendaid ="+a.getID();
-        sql = sql.replaceAll("$1", ""+a.getProfessor().getFunc().getID());
-        sql = sql.replaceAll("$2", ""+a.getAluno().getID());
-        sql = sql.replaceAll("$3", "" +a.getDataProva());
-        sql = sql.replaceAll("$4", "" + a.getHoraini());
-        sql = sql.replaceAll("$5", "" + a.getHorafim());
-        sql = sql.replaceAll("$6", "" + a.getLocal());
-        sql = sql.replaceAll("$7", "" + a.getSituacao_prova());
-        sql = sql.replaceAll("$8", "" + a.getNota());
-        
-        return Banco.getCon().manipular(sql);
-    }
-     
-     public boolean apagar(int cod){
-         
-         return Banco.getCon().manipular("delete from agendaprova where agendaid = "+cod);
-     }
-     
-    public Agenda get(int cod) {
-
-        Agenda agenda = null;
-        ResultSet rs = Banco.getCon().consultar("select * from agendaprova where agendaid=" + cod);
-        DALProfessor dalp = new DALProfessor();
-        DALAluno dala = new DALAluno();
-
-        try {
-
-            if (rs != null) {
-                if (rs.next()) {
-                    agenda = new Agenda(rs.getInt("agendaid"), dalp.get(rs.getInt("professorid")), dala.get(rs.getInt("alunoid")),
-                            rs.getDate("dataprova").toLocalDate(), rs.getString("horainicio"), rs.getString("horafim"), rs.getString("local"), rs.getString("situacaoprova").charAt(0), rs.getDouble("nota"));
-
-                }
-            }
-
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-
-        return agenda;
-    }
-
+  
     public ArrayList<Agenda> get(String filtro) {
 
        String sql="select * from agendaprova";
@@ -87,7 +43,7 @@ public class DALAgenda {
                  while(rs.next())
                 {
                     aux.add( new Agenda(rs.getInt("agendaid"), dalp.get(rs.getInt("professorid")) , dala.get(rs.getInt("alunoid")), 
-                            rs.getDate("dataprova").toLocalDate(), rs.getString("horaprova"), rs.getString("horafim"),rs.getString("local"), rs.getString("situacaoprova").charAt(0), rs.getDouble("nota")));
+                            rs.getDate("dataprova").toLocalDate(), rs.getString("horainicio"), rs.getString("horafim"),rs.getString("local"), rs.getString("situacaoprova").charAt(0), rs.getDouble("nota")));
                 }
             }
            
@@ -107,8 +63,9 @@ public class DALAgenda {
         try {
 
             Banco.getCon().getConnect().setAutoCommit(false);
-
-            ok = Banco.getCon().manipular("delete from agendaprova");
+          
+            if(!get("").isEmpty())
+                 ok = Banco.getCon().manipular("delete from agendaprova");
 
             if (ok) {
